functions/whatsapp-webhook/requirements.txt
flask
requests
psycopg2-binary
google-cloud-pubsub
gunicorn